requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.44
botocore==1.35.44
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.3
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.115.2
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.0
openai==1.52.2
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.1
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.40.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
websocket-client==1.8.0
wsproto==1.2.0
